+++ packages/graphics-engine/src/index.js
@@ -1,0 +1,9 @@
+export {
+  createGraphicsDocument,
+  createElement,
+  normalizeElement,
+  updateElement,
+  removeElement,
+  reorderElement
+} from "./document-model.js";
+export { DocumentHistory } from "./history.js";